refactor: コードの整理とイベント処理のクラスを新規作成、ファイル名を変更
